Move LineageHelper outside of antlr folder.
It's not connected to any antlr functionality.
