+++ src/index.ts
@@ -22,7 +22,7 @@ import { GroupByVisitor } from "./visitors/GroupByVisitor";
 
 // Antlr parser version
 export { default as antlr } from "./antlr";
-export { LineageHelper } from "./antlr/LineageHelper";
+export { LineageHelper } from "./LineageHelper";
 
 // Utils
 export { needToBeEscaped, printFilter };
